db+sqlite+pg: split GetAddressSecret into its own file
Move GetAddressSecret into addressstore_getsecret.go in each backend. It is a
thin read (one query plus row conversion), so per the wallet/internal/db
README it is a direct backend implementation with no ops interface.
