+++ wallet/internal/db/pg/addressstore_getsecret.go
@@ -1,0 +1,45 @@
+package pg
+
+import (
+	"context"
+	"database/sql"
+	"errors"
+	"fmt"
+
+	"github.com/btcsuite/btcwallet/wallet/internal/db"
+	"github.com/btcsuite/btcwallet/wallet/internal/sql/pg/sqlc"
+)
+
+// GetAddressSecret retrieves the encrypted secret information for an address.
+func (s *Store) GetAddressSecret(ctx context.Context,
+	query db.GetAddressSecretQuery) (*db.AddressSecret, error) {
+
+	var secret *db.AddressSecret
+
+	err := s.execRead(ctx, func(q *sqlc.Queries) error {
+		row, err := q.GetAddressSecret(
+			ctx, sqlc.GetAddressSecretParams{
+				WalletID: int64(query.WalletID),
+				ID:       int64(query.AddressID),
+			},
+		)
+		if errors.Is(err, sql.ErrNoRows) {
+			return fmt.Errorf(
+				"address secret for wallet %d address %d: %w",
+				query.WalletID, query.AddressID, db.ErrAddressNotFound)
+		}
+
+		if err != nil {
+			return fmt.Errorf("get address secret: %w", err)
+		}
+
+		secret, err = addressSecretRowToSecret(row)
+
+		return err
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	return secret, nil
+}
